Narrative Cell Update / Narrative v3.0.0-alpha-1

+++ scripts/install_narrative.sh
@@ -32,10 +32,14 @@
 # 8. Done!
 
 JUPYTER_NOTEBOOK_INSTALL_DIR=jupyter_notebook
-JUPYTER_NOTEBOOK_TAG=4.1.0
+# JUPYTER_NOTEBOOK_REPO=https://github.com/eapearson/notebook
+# JUPYTER_NOTEBOOK_TAG=4.2.0-kbase
 
-IPYWIDGETS_INSTALL_DIR=ipywidgets
-IPYWIDGETS_TAG=4.1.1
+JUPYTER_NOTEBOOK_REPO=https://github.com/jupyter/notebook
+JUPYTER_NOTEBOOK_TAG=4.2.1
+
+# IPYWIDGETS_INSTALL_DIR=ipywidgets
+# IPYWIDGETS_TAG=5.0.0
 
 PYTHON=python2.7
 
@@ -90,6 +94,7 @@ function make_activate_venv () {
 # -----------
 
 no_venv=0
+update_only=0
 while [ $# -gt 0 ]; do
     case $1 in
         -h | --help | -\?)
@@ -103,6 +108,10 @@ while [ $# -gt 0 ]; do
         -v | --virtualenv)
             make_activate_venv $2
             shift 2
+            ;;
+        -u | --update)
+            update_only=1
+            shift
             ;;
     esac
 done
@@ -126,33 +135,38 @@ then
   exit 1
 fi
 
-# Install external JavaScript code
-# --------------------
-cd $NARRATIVE_ROOT_DIR
-npm install >> ${logfile} 2>&1
-bower install --allow-root --config.interactive=false >> ${logfile} 2>&1
+if [ ! $update_only -eq 1 ]
+then
+    # Install external JavaScript code
+    # --------------------
+    cd $NARRATIVE_ROOT_DIR
+    npm install >> ${logfile} 2>&1
+    bower install --allow-root --config.interactive=false >> ${logfile} 2>&1
 
+    cd $VIRTUAL_ENV
+    # Install Jupyter code
+    # --------------------
+    # 1. Setup Jupyter Notebook inside virtualenv
+    log "Installing Jupyter notebook using $PYTHON"
+    console "Installing Jupyter notebook from directory '$JUPYTER_NOTEBOOK_INSTALL_DIR'"
 
-cd $VIRTUAL_ENV
-# Install Jupyter code
-# --------------------
-# 1. Setup Jupyter Notebook inside virtualenv
-log "Installing Jupyter notebook using $PYTHON"
-console "Installing Jupyter notebook from directory '$JUPYTER_NOTEBOOK_INSTALL_DIR'"
-git clone https://github.com/jupyter/notebook $JUPYTER_NOTEBOOK_INSTALL_DIR
-cd $JUPYTER_NOTEBOOK_INSTALL_DIR
-git checkout tags/$JUPYTER_NOTEBOOK_TAG
-pip install --pre -e . >> ${logfile} 2>&1
-cd ..
+    # This will clone the specified tag or branch in single-branch mode
+    git clone --branch $JUPYTER_NOTEBOOK_TAG --single-branch $JUPYTER_NOTEBOOK_REPO $JUPYTER_NOTEBOOK_INSTALL_DIR
+    cd $JUPYTER_NOTEBOOK_INSTALL_DIR
+    # git checkout tags/$JUPYTER_NOTEBOOK_TAG
+    pip install --pre -e . >> ${logfile} 2>&1
+    cd ..
 
-# Setup ipywidgets addon
-log "Installing ipywidgets using $PYTHON"
-console "Installing ipywidgets from directory 'ipywidgets'"
-# git clone https://github.com/ipython/ipywidgets
-# cd ipywidgets
-# git checkout tags/$IPYWIDGETS_TAG
-pip install ipywidgets==$IPYWIDGETS_TAG >> ${logfile} 2>&1
-# pip install -e . >> ${logfile} 2>&1
+    # Setup ipywidgets addon
+    log "Installing ipywidgets using $PYTHON"
+    console "Installing ipywidgets from directory 'ipywidgets'"
+    # git clone https://github.com/ipython/ipywidgets
+    # cd ipywidgets
+    # git checkout tags/$IPYWIDGETS_TAG
+    # pip install ipywidgets==$IPYWIDGETS_TAG >> ${logfile} 2>&1
+    pip install ipywidgets >> ${logfile} 2>&1
+    # pip install -e . >> ${logfile} 2>&1
+fi
 
 # Install Narrative code
 # ----------------------
@@ -169,33 +183,43 @@ ${PYTHON} setup.py install >> ${logfile} 2>&1
 log "Done installing biokbase."
 cd $NARRATIVE_ROOT_DIR
 
-# Install KBase data_api package
-# ------------------------------
-git clone https://github.com/kbase/data_api -b develop
-cd data_api
-pip install -r requirements.txt
-$PYTHON setup.py install >> ${logfile} 2>&1
-cd ..
-rm -rf data_api
+if [ ! $update_only -eq 1 ]
+then
+    # Install KBase data_api package
+    # ------------------------------
+    git clone https://github.com/kbase/data_api -b develop
+    cd data_api
+    pip install -r requirements.txt
+    $PYTHON setup.py install >> ${logfile} 2>&1
+    cd ..
+    rm -rf data_api
 
-# Setup jupyter_narrative script
-# ------------------------------
-console "Installing scripts"
-i=0
-while read s
-    do
-        echo $s
-        if [ $i = 0 ]
-            then
-            echo d=`pwd`
-            echo e=$(dirname `which python`)
-            i=1
-        fi
-done < $SCRIPT_TEMPLATE > $SCRIPT_TGT
-d=$(dirname `which python`)
-chmod 0755 $SCRIPT_TGT
-log "Putting new $SCRIPT_TGT command under $d"
-/bin/mv $SCRIPT_TGT $d
-log "Done installing scripts"
+
+    # Setup jupyter_narrative script
+    # ------------------------------
+    console "Installing scripts"
+    i=0
+    while read s
+        do
+            echo $s
+            if [ $i = 0 ]
+                then
+                echo d=`pwd`
+                echo e=$(dirname `which python`)
+                i=1
+            fi
+    done < $SCRIPT_TEMPLATE > $SCRIPT_TGT
+    d=$(dirname `which python`)
+    chmod 0755 $SCRIPT_TGT
+    log "Putting new $SCRIPT_TGT command under $d"
+    /bin/mv $SCRIPT_TGT $d
+    log "Done installing scripts"
+
+    log "oh, wait, one more thing...installing nbextensions"
+    cd nbextensions
+    sh install.sh
+    cd ../..
+    log "now, done."
+fi
 
 console "Done. Run the narrative from your virtual environment $VIRTUAL_ENV with the command: $SCRIPT_TGT"
